move load handler to pref.py
