+++ release/scripts/startup/abler/__init__.py
@@ -17,19 +17,6 @@
 # ##### END GPL LICENSE BLOCK #####
 
 
-from . import pref
-from . import render_control
-from . import layer_control
-from . import view_control
-from . import shadow_control
-from . import image_adjustment
-from . import face_control
-from . import edge_control
-from . import scene_control
-from . import general
-from . import credential_modal
-from . import custom_properties
-import bpy
 bl_info = {
     "name": "ACON3D Panel",
     "description": "",
@@ -45,6 +32,20 @@ bl_info = {
 
 
 # Main imports
+import bpy
+
+from . import pref
+from . import render_control
+from . import layer_control
+from . import view_control
+from . import shadow_control
+from . import image_adjustment
+from . import face_control
+from . import edge_control
+from . import scene_control
+from . import general
+from . import credential_modal
+from . import custom_properties
 
 
 # =========================================================================
